! refactor converters to abstract traverse and function removes duplicate DOM traversal code and allows converter classes to focus on  conversion and not the abstract traversal.  Moved to own namespace.  Removed several crusty code areas that were dealing with broken emails etc.
